Move legacy code for VCL a11y into header helper
