+++ src/platform/site-wide/header/components/App/index.js
@@ -5,7 +5,10 @@ import { connect } from 'react-redux';
 // Relative imports.
 import Header from '../Header';
 import LegacyHeader from '../LegacyHeader';
-import { addOverlayTriggers } from '../../helpers';
+import {
+  addFocusBehaviorToCrisisLineModal,
+  addOverlayTriggers,
+} from '../../helpers';
 
 const MOBILE_BREAKPOINT_PX = 768;
 
@@ -31,6 +34,9 @@ export const App = ({
 
     // Creates trigger function that opens/closes mobile menu, mobile search menu, and the Veterans Crisis Line.
     addOverlayTriggers();
+
+    // Adds focus behavior to crisis line modal.
+    addFocusBehaviorToCrisisLineModal();
 
     // Clear listener.
     return () => window.removeEventListener('resize', deriveIsDesktop);
